Refactor: restructure ray type hierarchy, expand fields
This essentially employs a fatty struct system, at least to the extent
that the paradigm / model / framework makes sense in Julia. This was
already being done with the System struct anyway, this simply continues
the approach.

Having an unconventional plural ray type was unergonomic and so was
having to take views when unnecessary and pass separate refractive index
vectors. Now Lens stores the refractive indices, moving the work to
internal instance construction rather than API methods.

Here there is some amount of redundancy, but I think this design makes
more sense. We now parametrize rays by type, and the distinction between
optical and physical rays is no longer necessary as the slopes for the
latter can easily be retrieved using a convenience method; having
separate OpticalRay and ParaxialRay types was confusing given that
OpticalRays are also paraxial rays.