`construct` should be a `Lens` constructor
Including the refractive indices in the lens as part of the previous
refactor, aside from being more intuitive and pragmatic, opened up the
ability to utilize the obvious Lens method for construction as the
increased arity allows for discrimination between surface and lens
matrices by using multiple dispatch to delegate between direct lens
matrix + n vector input and an input surface matrix.

Also, while I was at it I noticed taking that view was unnecessary as
instantiating the lens will allocate an array anyway so taking a slice
is equivalent.