Fix for afocal systems, include BFD in total system length calculation
Full solving support for afocal and telecentric systems is limited at
the moment due to the propagating Infs and NaNs, but this should improve
the plots.

As for the focal distance, the total system length governs the object
space ray distance so including it should improve the plot scale; right
now it's taking 10% of this total length for that distance.

Also, I noticed the calculation for the EBFD could be simplified.