`rayplot`: remove the axis labels
Since this is calling DataInspector which uses the standard Cartesian
coordinate labels this was a bit confusing so instead of formatting a
custom inspector_label it's better to just remove the axis labels
completely. It doesn't make much sense to specify an axis convention
when it's arbitrary and the system being modelled could have a different
configuration; the units are already not being included because the
scale is input dependent and since this is a simple analysis tool with
the plot serving as a visual aid such labels shouldn't be included.

current_axis().xlabel etc. could be used if a better presentation is
needed.