Include the physical paraxial angles in the Ray struct
Again following a fatty struct design with redundancy; I found that I
needed this while implementing the Seidel sums and calling the slopes
function wasn't very satisfactory. It's useful to not only have this
precomputed, but also in tabular form as an array.

The extended refractive index vector is also included for ease in length
matched calculations.