v0.3.0
Once I get this package to version 0.4.0 or so I'll restructure the
package into separate files.

Next version will include drawing the lens surfaces on the plot which
should be fairly straight-forward and bringing MakieCore in as a light
dependency in order to control compatibility and allow defining a plot
recipe.

Long term roadmap includes adding constraint based system solving and
aberration optimization, but that's a long ways away -- planned for
version 1.0.