Have raytraces return a `Ray`, store the `z` coordinate in it
This refactor moves some of the computation toward ray construction as
having a full set of coordinates over the plane stored in the ray struct
is useful. This also moves toward being able to plot any general ray or
ray bundle. The next commit will add the rest needed for that, as what
is missing are general raytrace methods given an objective distance and
optionally a System.