`raytrace(system, etc.)`: avoid actually raytracing
So since we already have the marginal and chief rays here we can
actually avoid the whole rigmarole of a full retrace due to the
linearity of the system. In particular, we can take advantage of the
fact that any ray can be expressed as a linear combination of any two
linearly indepedent rays; the system is a two-dimensional vector space
spanned by two such rays {(y1, u1), (y2, u2)} and thus the marginal &
principal rays form a basis -- this is essentially the principle of
optical invariance in the paraxial regime.

So we can just solve L * K = L_s at the pupil where L is the ray matrix
[y ȳ; nu nū] with its determinant being the Lagrange invariant, L_s is
the new ray matrix, and the change of basis operator K ≡ [a b; c d]
contains the coefficients to be used in computing the new ray values at
each surface by applying this linear transformation.

In this particular case K simplifies to [1 0; α β].