Vignetting: emit NaNs if aperture size is limited by the marginal ray
This also includes the size of this limit at which point the aperture
becomes the stop. The way it was set before could be confusing, this
makes things clearer and easier to analyze.

Also, the formatting on the column label printing is kind of wonky, but
for now I'm not going to spend any more time on it as it's a minor
detail compared to what still needs to be implemented.