Simplify calculation of the chief ray
Following the approach outlined in commit fee946d, we can again use
the linear property of the system to determine the chief ray.

Before this a convoluted approach of tracing from the stop to image
space and then reverse tracing from the stop and concatenating was used.

Now, a preliminary ray (y2, nu2) is launched and then transformed into
the principal ray. Note that this preliminary ray is chosen such that it
is linearly indepedent of the marginal ray (y, 0.0).

What we're actually solving is:

L * [α, β] = [0.0, nū_stop]

This yields [α, β] = (H / I) * [-y2_stop / y_stop, 1] where H is the
Lagrange invariant and I is the optical invariant of the
(marginal, preliminary) system; because we're initially tracing
(0.0, 1.0) this invariant nu2 * y - nu * y2 simplifies to just the
marginal ray height at the objective lens which is also the marginal ray
height at the entrance pupil under collimation; the ratio of invariants
winds up being the object space chief ray slope. The chief ray is then
the field scaled superposition of the preliminary ray and the marginal
ray scaled by the negative stop height ratio.