Delete a broken `raypoints` method
Plotting single rays is not supported at the moment due to the way the
coordinates are stored as the plotting mechanism depends on both the
marginal and chief ray information for the object and & image space
extensions. For any ray calling lines! and plotting manually is of
course possible, but some work has to be done by the user in extracting
the desired points and extending appropriately if overlaying it on top
of an existing plot is desired. For now I think the plotting does the
job it's supposed to and time is better spent working on other features
or fixing things.