v0.5.0
Plan for upcoming versions:

Version 0.6:
wave-fan and ray-fan plots using the aberration coefficients

Version 0.7:
modulation transfer function analysis

Version 0.8:
constraint based optimization

Version 0.9:
real / full / non-paraxial / non-linear ray tracing and higher order
aberration analysis

Version 0.10:
spot diagrams

All depending on how straight-forward they are to implement. If those
features work well I'll release version 1.0.

As far as improvements goes, the vignetting tool is currently very slow
right now, I suspect because of all of the searching and specialized
printing, but this isn't meant to be called in loops or anything
(although this is done in the tests for verification and it does take a
while) so performance isn't that big of a deal for this as it's meant
to be a few shot analysis function; still, this can be improved.