Implement the aberration plots
A bit of boilerplate, but this is the most direct way to do it. I'm not
bothering with a recipe for these because with the exception of the
distortion plot these aren't simple plots, they're compound plots in
which we're setting axes and laying multiple plots. The simplest
implementation here are direct functions especially since we're not
specifying any extra attributes or messing with themes, so we can just
pass the kwargs around.