`aberrations`: use absolute value of the chief ray refraction invariant
Tracing a system with an image height of opposite sign will flip the
sign for the odd aberrations and since we're normalizing the field
assuming a positive sign this needs to be taken into account. Rather
than storing the height or its sign in the aberration data and
conditionally signing the field this is the easiest way to deal with it.