Include the stop in the `System`, rework `vignetting` dispatch
Again it's better to simply store everything in the struct as it makes
everything easier to use. Note the surfaces cannot be stored as the
solving mechanism operates directly on a `Lens` since lenses can be
directly specified without having to construct one from a surface matrix
so it cannot have knowledge of the previous information; having the
surfaces and the `System` separate makes sense though and it also allows
intuitive dispatch techniques.

As for vignetting, the external API using a singleton type for dispatch
could be confusing and `FOV` shouldn't be a reserved exported name.

This rework does remove the ability to vary the aperture sizes using the
first method, but this is not necessary as the matrix carries all of the
necessary information for this type of analysis. It's more intuitive to
specify just a system, which includes the field of view, in order to
return the supported aperture sizes, and a system + apertures in order
to extract the supported fields. Internal dispatch with `Lens` is fine
for now, but this is all subject to change.