Add spot size diagrams
This is probably inefficient and is subject to a refactor, but it seems
to work for now. Note because we allow the lattice finesse to vary the
transverse ray errors need to be recomputed to calculate the RMS spot
size; this is because the centroid only appears in the lattice if an odd
number for k is chosen -- if it is then it's possible to just index the
lattice, e.g. lattice[:,div(k, 2)] then stacking it and viewing through
the appropriate row will extract the tangential or sagittal vector.

In any case, performance does not appear to be an issue for the default.