Tests: use the image space paraxial numerical aperture for the scaling
Also, clarify the accompanying comment.

Transverse ray errors are defined in terms of the exit pupil and not the
entrance pupil so using the image space slope here is more accurate and
since we have this value, especially the one from their trace, it's
better to just use that.

As for the comment, I think I finally understand why the Seidel
coefficients are scaled the way they are in order to convert them to
wavefront expansion coefficients. They refer to the ray error expansion
coefficients, not the wave errors, under the supposed historical
convention of scaling by the f-number. The addition of the first-order
chromatic contributions in terms of a Seidel-like sum must have come
later and use the more modern convention.