Implement a provisional real ray trace
This will remain an internal function until I can verify it and
integrate it with the rest of the package. Dispatching with Real isn't
necessary right now, but it's how I plan to eventually bring it in, so
I'm leaving it like that.

Also, I noticed that invoking size for the number of columns of the
paraxial raytrace isn't necessary as it should always have two columns,
so I went ahead and changed that; probably should be its own commit, but
whatever.

As for the formulas, I derived them myself because the geometry is
simple and it's just a direct application of Snell's Law, but I'm still
a bit unsure as to whether they're actually correct, and there's
probably a simplified and more efficient version out there.

The validation in the tests uses residual asymptotic bounding on the
truncation error for the Maclaurin expansion of sine, and I figured the
total error would propagate roughly as the number of operations up to
each surface which results in a triangular number in the end. The test
passes so I think the implementation is right at least to some degree.