Improvements to real ray tracing
This now takes the surface sagitta into account on transfer. I've made
some attempt to fully account for the circular geometry and TIR, but
again I'm not exactly sure this is correct as this is all self-derived.

The function is now exported for my own personal convenience, but it
will remain undocumented for now, at least until I can verify it; it's
not likely to be called on accident given the type dispatch, which, by
the way, is subject to change as it's awkward to use a standard Julia
type for this, but I think the naming makes sense and since the
parameter is not used for any of the fields it won't affect performance.

ParaxialRay has been renamed to bring it into the convention we're
already using for the other rays.

As for the approximation error - I've noticed that for any real kind of
system even the third order aberration values are substantially off.
Assuming the real ray tracing implementation is correct if you trace a
collimated marginal at the height of the paraxially derived entrance
pupil the only monochromatic aberration that should be present is
spherical aberration and transferring the real ray to the focal point
yields the transverse ray error which can be compared with the callable
RayError object evaluated at the edge of the pupil with zero obliquity;
and for the example Tessar lens the relative error is about 66% - about
60% if you try to get the real ray to fill the aperture stop.

I suppose this makes sense, given that the upper bound on the
approximation error for third order aberrations scales as a quintic
raising the height by an order of magnitude should yield rather large
discrepancies and since we're tracing right at the edge the full weight
of the coefficients is preserved so for fast lenses there should be
quite a deviation; I suppose the fifth and seventh order aberration
estimates should cover a decent portion of the error.