Improve determination of z-coordinate end points at the Ray struct
This makes it so that non-characteristic rays don't have a meaningless
end coordinate and now the initial point is determined in the inner
constructor rather than at raypoints. This also effectively removes
support for plotting afocal systems, but full support for that wasn't
there anyway, so rather than partially support it with redundant code
it's better to just remove it; afocal systems are usually modelled by
placing a lens after the system anyway and looking at it that way.