Find the real marginal ray
This employs a binary search method which is logarithmic in time
complexity, but only linear in convergence as the error is halved after
each step. Newton's method in comparison has quadratic convergence, but
at the moment I don't want to deal with possible instabilities of the
finite differences at the stop and the algorithm not converging.

The drawback to using bisection here is the need for the initial
overshoot or undershoot in order to find a suitable bound to bracket,
but this interval is not difficult to find given the paraxial estimate.
I can always use this as a fallback method if I employ a more
sophisticated approach later on.