Create a Prescription type
This mostly provides a convenience constructor for now, but the reason
I'm adding it, even though the raw float matrix is more primitive within
our scheme (compared to Lenses which are constructed from surface data)
is to future proof things so I don't have to deal with this later.

I expect to return custom wrapped surface prescriptions when I add
constraint based optimization; the idea there is that some properties of
the system would be fixed while allowing some of the prescription
parameters to vary -- by passing along indices or what not -- and then
some form of autodiff would be performed minimizing aberrations,
possibly with some shape factor heuristics and then the mechanism would
provide the new lens prescription. This is still a long ways out, but
I'm putting this type system in place so it's there.

At the moment this simply allows Prescription(; R, t, n) surface
construction from named vectors.