Add a tranverse spherical aberration function for real ray errors
There's some redundant code with the computations at caustic, but that's
fine; if this type of extension pattern becomes more common I'll add
functions for it, but I'm used to transferring explicitly so I'm more
comfortable writing it in-line than having to pass arguments and
remember their position.

Anyway, this now allows a full specification of the actual amount of
spherical aberration produced by an optical system given its
prescription; the returned vectors here can be used in conjunction with
my Zernike package to compute its wavefront expansion. The only thing
that's missing for a full analysis of that aberration is the modulation
transfer function which I initially planned to add to this package, but
realized it's a much better fit for Zernike so I will end up adding it
there.