Add some transfer functions for real ray tracing
On second thought I decided to add these so that I only have to maintain
one thing. It does clean up the clutter at the call sites and makes it
easier to read, but the drawback is that the calculations are now
scattered rather than collected all in one place -- having it structured
in an imperative, linear fashion is sometimes useful in catching bugs
due to an incorrect implementation as the logic is all layed out in a
straight-forward fashion; but of course if there is a bug this is easy
to fix across the entire package with a drop-in using a functional
approach; analyzing it piecewise is also sometimes useful.

Forcing it through multiple dispatch has the advantage of catching those
types of errors as well: for example, marginal rays are indexed
differently compared to tangential rays as they carry with them image
space axial coordinate metadata and so their extension from a surface is
different -- having separate methods for this makes sure the proper
computation is carried out.

In any case, reuse is good here in case I change the way the coordinate
system works.