Add example code for fitting real TSA to Zernike polynomials
This is useful to me, but right now maybe it doesn't belong so much as
part of the package proper so it'll be treated as extra / miscellaneous.

Eventually this should probably be part of the documentation using
example blocks and what not, but for now it will remain a script.