Find the real chief ray
This uses Newton's method instead of the binary search employed for the
marginal ray. This technique avoids the need to provide initial bounds;
the bracketing approach used for the marginal ray appears to not be a
problem as reasoning about perturbations around the stop with respect to
changes in launch height is clear and we can just shift past the zero
with some amount of certainty, but for the chief ray it's not so obvious
to me as we're varying the angle and I find reasoning about it this way
much easier. This benefits from an improved convergence rate as well.

What should be added now is a parameterization of RayBasis as an
extension for real rays and an analogous function to TSA which varies
field angle; that type of function should be able to fully quantify all
aberrations at least in the tangential and sagittal directions at the
image plane given the asymmetries present between the aberrations for
those two orthogonal planes; I will have to think about that some more,
but it seems like it should be possible.