`TSA`: trace the real chief ray, compute & return the exit pupil heights
So given the system here cannot be treated as linear unlike the paraxial
case we can't just assume invariance and a constant entrance to exit
pupil mapping; in the paraxial case the pupil ratio is characterized by
the angular magnification and once the values are normalized the spacing
should remain the same, but here I don't think we can do that. Given
that wave aberrations are defined with respect to the reference sphere
centered at the exit pupil and the power series expansion uses the exit
pupil coordinates, then using the entrance pupil heights is incorrect.