Fix real ray tracing, again -- solve the actual surface sagitta
So apparently I had calculated an approximation since I didn't take the
full geometry into account. The ray was transferring across the initial
sagitta and then the sagitta was being recomputed, but neither coincide
with the actual surface. I discovered this trying to verify the real
chief ray trace and finding that the values were off by a slight amount.

There might be a simpler way to deal with this, but my approach here is
to solve for the point of intersection and compute the sagitta using the
axial angle.