v0.6.0
This is a re-release as I thought it sensible to include all of the real
ray tracing additions, improvements, and fixes in the same version.

I had planned to add optical transfer functions, but that was much
better suited for my Zernike package so I added them over there.

I added real ray tracing sooner than expected, but it's not fully
implemented yet and it also needs to be optimized. Eventually a three-
dimensional spatial trace using direction cosines should be added, and
then we can just use that for full aberration analysis. Or maybe there's
a way around that using some clever approach, at the moment I don't
know.

What's still missing for version 1.0:

* quantification of higher order aberrations

* system constraint optimization

* support for aspheric surfaces

I've begun the work on aspheric surfaces at a feature branch, but it
appears to be more difficult than I initially anticipated, so maybe that
will have to wait.

As for the other two features, there does not appear to be analytic
forms for higher order terms of the aberration expansions like there are
for the third-order Seidel coefficients, so maybe full non-paraxial
tracing plus fitting needs to be employed for that; if that's the case
then the surface solving / optimization should come first as the loss
function probably needs to be relatively quick for that to work well,
but I'm a bit unsure as to whether optimizing for third-order
aberrations is a good approach given how inaccurate they can be and how
real systems are balanced using higher order terms. In any case, it's
probably a good starting point and it's a good idea to put the
optimization system in place regardless of what is actually minimized.