Add primitive support for aspheric surfaces
This is an initial sketch / draft and is not working yet.

The scaffolding is in place to work with these types of surfaces, but
the computation needs to be carried out correctly. It seems to me that
generalizing the surface profile is not naively possible as parabolic
surfaces need to be treated specially as that 1 + K factor vanishes. In
addition the surface normal / tilt angle θ needs to be handled depending
on the conic constant as it will vary.

Also, this only considers quadrics and disregards aspherics defined in
terms of polynomials.