Add the constraint optimization algorithm
Using a straight penalty is the most basic and hassle free approach, and
from my limited testing it seems to work well, at least for what the
tool is meant for.

I tried implementing a more sophisticated augmented Lagrangian approach
using adaptive multipliers, but it was more trouble than it was worth
and in the end just hard-coding the penalty and not bothering with an
outer loop, updating parameters, or scaling constraints worked just as
well if you target appropriately.

I also looked at JuMP but that's a domain specific language and a heavy
dependency compared to something like Optim, so it seemed like overkill.
It also adds quite a bit of complexity and our functions need to support
dual numbers for automatic differentiation; structs need to also be
parameterized to support this.

The default algorithm for Optim also works well and is quite fast as
it's derivative free.

Right now the goal is to keep things simple and light, as the focus is
on providing all of the functionality necessary to analyze a system
quickly.