Add a test for the new optimizer
The shape factor test actually fails if the lens is stopped down as that
drives down the aberration values and the function seems to struggle
finding the optimal solution when values are relatively close to zero
already.

If the power is increased the test will also fail as that blows up the
value instead. Maybe there's a scaling trick that can mitigate this.
The SI coefficient, for example, depends quartically on pupil position
and cubically on power, so it can be fragile.

The solution also appears to depend on the sign of the curvature when
the radius is large.

Also, the optimizer doesn't handle plano surfaces yet, so any Infs will
break it.