Implement the Augmented Lagrangian optimization method
As commented, lots of arbitrariness, but results seem to improve. I've
not rigorously tested this, however. I'm also not sure I'm implementing
this correctly, either. For the example Tessar lens this optimizes quite
well though and all of the third order aberrations vanish -- only the
Petzval contribution (which is a first order aberration) remains. The
Cooke triplet in the tests is well corrected for spherical, balanced by
the fifth and seventh order terms, so optimizing that requires
targetting specific aberrations. Either way, at the expense of added
complexity I think this is a better approach.