`optimize`: take constraint scale into account
I think at this point with this many closures, the involved logic, how
large the function is getting, and the delicate balanced needed to
retain type stability I won't be adding any more complexity.

This was meant to be simple, but I couldn't help but try and improve it.

I think anything more than this would require supporting full autodiff
and bringing a non-linear solver into the mix, a package like JuMP would
be the best bet.

Improving this specific algorithm probably involves tweaking the
constants.

Improving the quality of the objective function is the next natural step
as it's currently only taking the third order coefficients into account.