`optimize`: explicitly specify type parameter for penalty wrapper
Ref(x::T) lowers to RefValue{T}(x), but I prefer being explicit here
given the point of why it's being used in the first place. Also, the
comment was adjusted because it made it seem like heap allocation is
what is being avoided, but of course Ref is also heap allocated -- the
point is that the compiler isn't able to infer the type, Core.Box is
type agnostic, and using a pointer-like Ref with a concrete type could
improve performance. We could also use `let` blocks, but the multiple
levels of nesting with closures could make the source code here more
difficult to follow and given that it's being shared across multiple
scopes using a container is better.