Finish implementing aspheric conic sections
Again self-derived as it's faster than looking for the relevant formulas
in a book or online and working from first principles helps me
understand things better and keep things simple. Of course that means I
might not be implementing it correctly so I need to make sure I write
good tests; right now testing aspherics is very basic, it's mostly
a simple safeguard / check. Drawing the caustic for the reflector
shows all rays focusing at the paraxial plane as expected; I'm sure
there are other things I could test, but right now I can't think of any
so I might come back to this once I finish implementing everything else.

Anyway, the branch for tilt is in fact faster than just going through
the general computation including the conic constant, about 7% faster,
which for a single trace is fractions of a microsecond, but if we're
tracing large numbers of real rays it could make a difference.