Implement support for aspheric surface polynomial functions
This adds a bit of complexity, but is much more flexible and general.
I initially thought of just adding support for the specification of
a vector for the coefficients c_k of the series:

Sum from k=2 to n of c_k * y^(2k) =  Ay^4 + By^6 + Cy^8 + ... + Zy^n

which would make the derivatives simple and exact, but allowing any
general function means you can use Q-polynomials, Chebyshev polynomials,
and really anything you like; although the assumption of rotational
symmetry is still there as the x-coordinate is not used, so the surfaces
cannot be set entirely freeform in that sense, this still provides more
freedom at the cost of complexity.

There is the added assumption that the provided functions must support
complex evaluation in order for the complex step derivatives to work for
the surface normal computation, which are used to eliminate floating
point round-off error compared to using finite differences, but that's
fine as most standard library Julia functions support this and it's
expected that the functions will be relatively simple and polynomial in
form. Of course truncation error is still there, which is the drawback
here compared to just supporting the coefficients, but I think that's a
good trade-off.