Tests: fix the Petzval curvature test
The longitudinal Petzval aberration was being scaled using the book's
value for the marginal slope, but our Petzval coefficient; in order to
properly compare the scaling needs to be consistent as the trace values
differ by a slight amount. Note that -1/n′u′ = 2N and the longitudinal
aberration picks up an extra factor of this along with the value of the
degree imparted by the derivative and the negative sign for the ray
aberration thus yielding -8N ^ 2.

As for the Petzval curvature defined in terms of the sum of the surface
powers the equation being used was incorrect. I had used an equation
defined in a different book and I decided to check the result
mathematically and it turns out that the Petzval sum has to also be
divided by the preceding index in order for it to be equivalently
related to the Seidel sum term being used to calculate the curvature.

The equation in that book assumes surfaces bounded by air on one side
and so one of these indices is always equal to unity and thus one of the
terms in the divisor vanishes.

The tests were previously passing for two reasons:

- the image space marginal ray slope from our trace is very close to
  the one provided in the book and the tolerance was large enough

- the Cooke triplet has only surfaces bounded by air

However, if you have a system with buried surfaces, doublets, cemented
surfaces, etc. the relation breaks down. This is in fact the case with
the example Tessar lens and once these refractive indices are taken into
account the comparison succeeds.

The second test doesn't need a tolerance specified as what it's really
comparing is the implementation of the Seidel sum to a mathematically
equivalent formula for the Petzval curvature.