Start implementing full 3D real ray tracing; add vector refraction
This begins a series of commits that should culminate with version 1.0.

What will be added:

* Pupil sampling; this will loop using polar coordinates and compute the
  ray errors from the paraxial point in both directions.

* Full tracing for any conjugate distances; this will involve solving
  the axial angles as they are coupled to the pupil coordinates given a
  RayBasis which has the object height and distance encoded within it.

Full tracing for a planar wavefront should be straight-forward as the
angle V is zero and the angle U is constant.

This current commit adds the base tools for implementing the scheme.

I could have gone with the usual approach of using the axial angles
and computing the amount of refraction using the trigonometric relation
in both directions, but I think moving towards a vector approach is
sensical as refraction can be computed in one go, and I also wanted to
understand it; the vectorial Snell's Law equation is derived by
decomposing the wave-vector into parallel and perpendicular components
and using the tangential continuity condition at the interface plus
projections for the normal direction along with some sign and
discriminant considerations to account for reflection.

The drawback is the need to test it, as the approach used for tangential
tracing has been verified to work, but this technique might have a bug
in it, especially with regards to signs and its attempt to be general.