Implement the `full_trace` pupil sampling function
This is not finished yet, the off-axis fields need to be verified and
everything still needs to be checked for implementation errors. The
commented lines are the naive implementation without regards for
symmetry; those will be removed once I verify everything is correct --
the idea is that for a rotationally symmetric / axisymmetric system only
half the pupil needs to sampled as the errors should be symmetric with
respect to the meridional plane, which is also why only the object in
the meridional plane is considered in the trace.

Ideally real ray trace spot diagrams should be used to compare this
implementation so I'll have to look for those once I finish adding
everything.

What's left in the series once this is complete:

- tests: check vector raytrace against the real tangential rays
  - check against a reflector + lens system for sign consistency
- add a pupil_shift function which realigns the matrices according to
  the other angle convention
- add a real spot_size method which redirects to a scatter plot of these
  vectorized ray error matrices
- document

Also, I am debating as to whether these 3D trace methods should return
the per-surface results, but that just adds more complexity and the need
to allocate when the quantites of interest are the final coordinates.
I think the tangential trace method does a good enough job if that data
is required.

As for the RayBasis finite distance consideration, this is the easiest
way to implement it. Sampling for planar ray fields is, as mentioned in
the previous commit, direct and easy to reason about, but for
non-collimated rays additional calculations need to be performed as the
strike angles now vary.