Utilize a `RealRayError` struct, add the `wavegrad` function
wavegrad is a transformation function (scaling + rotation) which returns
the wavefront error partial derivatives given the ray error matrices.

Note I'm not adapting it for straight spherical aberration fitting as
that's mostly meant to be a reference and tutorial. I have changed the
calculation of the wavefront derivative there though as the reference
sphere is defined using the paraxial quantity.