Fix wavegrad, switch to naive polar trace, extract surfaces with a view
This will mostly wind up as a squash commit.

Rotating the matrix is strictly incorrect, the quadrants need to be
rearranged, but for now I am going abstain from manipulating it.

The surface matrix can vary in the number of columns as aspherics
specify the conic constant so that is essentially stripped and passed as
a keyword argument instead.

The full tracing is incorrect at the moment:

The spot diagram is not uniform for the on-axis trace; only spherical
aberration should be present there and it should look symmetric.

When the wavefront gradient is fit to Zernike polynomials there's also a
significant amount of trefoil present which should not be the case. The
derivatives should be comatic with some tilt which result in a symmetric
wavefront error when converted. There's an asymmetry being introduced
somewhere in the trace. There's something I'm not quite understanding
here.