Fix the full trace by taking into account 3D spherical geometry
I realized that the sharp quadrilateral-like features showing up in the
spot diagrams as well as the oscillations near the extrema of the pupil
edge slices for H = 0 were being caused by a field dependent
magnification which shouldn't occur for an on-axis field meaning the
surface sagitta and gradient calculations were probably incorrect.

Note that I did not rigorously re-derive these relations, but rather
used intuition to account for the extra coordinate so they might be
incorrect; I'll mathematically recheck everything and adjust the
computations if they are later on during the verification phase.