`full_trace`: sample the disk uniformly, use the symmetry optimization
By sampling the radius non-uniformly the density of rays over the pupil
becomes uniform; otherwise the points become distributed in such a way
that the rays become concentrated near the center of the entrance pupil
and relatively more sparse near the edges, which is not what we're
interested in as the aberrations scale with pupil position.

There are other mappings that offset this, but this sampling strategy is
simple and intuitive and good enough for now. From what I can understand
this area bias is compensated for by ensuring that the square root of
the determinant of the metric in polar coordinates is invariant and the
transformation which preserves this quantity is √ρ.

The ray error struct now holds the input coordinates so that any fitting
procedure can realign / reorient the wave error pupil map according to
the desired angular convention.