`full_trace`: adjust initial launch coordinates, paraxial image height
This should take into account the asymmetry involved between the
tangential and sagittal directions. Since the initial trace inputs refer
to the objective rather than the entrance pupil the points need to be
adjusted so that the bounds are correct.

Also, since the real chief ray is defined from the edge of the sensor /
detector / specifed image height an extra paraxial trace needs to be
performed in order to find the image height in relation to the real
field of view.

The way both of these things are done is subject to change, at the
moment I'm quite unsure as to whether I'm doing any of this correctly.

The on-axis trace appears to be correct, but the geometry there is much
simpler; once an off-axis field tilt is introduced and the rays are no
longer collimated (RayBasis inputs) then the initial conditions become
much more difficult to reason about.