Vector refraction: don't modify signs
Reflective systems were producing incorrect results so I re-derived the
equations for both refraction and reflection from scratch; the equations
were correct for refractive systems, but I had tried to generalize the
formulation to account for reflective ones and had actually produced an
incorrect result. It turns out that encoding direction in the refractive
indices works out without having to modify the equation -- transfer is
taken into account with signed distances and the sign of the third
component of the direction vector doesn't actually matter.

Moreover, because this scheme is only used to determine the final
positional coordinates rather than a surface-by-surface trace, and the
optical axis z-coordinate isn't returned, using a formulation which in
effect unfolds a reflective system in the mirror sense seems to work.

I have not rigorously tested this, neither have I proved it
mathematically in full detail, but it makes sense at some intuitive
level, and for refractive systems this doesn't change anything.
Reflective systems appear to work now -- I'll add a basic test in the
next commit -- and I can always revisit this after I finish everything
else.