`full_trace`: attempt to find the pupil bounds for off-axis fields
This is more difficult than I initially thought. Apparently there isn't
a bijective pupil mapping for real rays as linearity no longer holds so
some rays if scaled uniformly will in fact miss the aperture stop and if
these are included the ray errors scale disproportionately distorting
the spot diagram. Since the objective is to have the ray errors
accounted for in a polar format the problem becomes even more difficult
because now even if these edge rays are found the spacing between them
and the chief ray is no longer symmetric. I think there might be some
way of projecting the pupil with the chief ray angle and setting down
some approximate bounds with the marginal ray height, but for now I'm
going to commit this and come back to it later.